+++ deps/docker_solcmc_updated
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+docker run -it --rm --user $(id -u):$(id -g) -v $(pwd):/smtch --workdir /smtch usiverify/testgen:latest /bin/sh run_solcmc $@
