--- lib/run_solcmc
@@ -1,5 +1,0 @@
-dir=`pwd`
-dir="$dir/$1"
-shift
-cd /home/solc-js/
-ts-node run.ts "$dir" $@
